requirements.txt
django==4.0.1
djangorestframework==3.13.1
django-filter==21.1
markdown==3.3.6
psycopg2-binary==2.9.3
pillow==9.0.0
crispy-bootstrap5==0.6

pytest==6.2.5
pytest-django==4.5.2

ipython>=8.0.1
flake8>=4.0.1
black>=21.12b0